small adjust

+++ app/controllers/cart.controller.js
@@ -1,5 +1,6 @@
 const db = require("../models");
 const asyncHandler = require("express-async-handler");
+const { ObjectId } = require("mongodb");
 
 const Cart = db.cart;
 const User = db.user;
@@ -9,6 +10,7 @@ const Product = db.product;
 
 const addCart = asyncHandler(async(req, res) => {
     const { _id, cart } = req.body;
+    console.log("🚀 ~ file: cart.controller.js:12 ~ addCart ~ cart:", cart)
     try {
         let products = [];
         const user = await User.findById(_id);
@@ -20,6 +22,7 @@ const addCart = asyncHandler(async(req, res) => {
         for (let i = 0; i < cart.length; i++) {
             let object = {};
             object.product = cart[i]._id;
+            console.log("🚀 ~ file: cart.controller.js:24 ~ addCart ~ cart[i]._id:", cart[i]._id)
             object.count = cart[i].count;
             let getPrice = await Product.findById(cart[i]._id).select("price").exec();
             object.price = getPrice.price;
@@ -41,10 +44,10 @@ const addCart = asyncHandler(async(req, res) => {
 });
 
 const getUserCart = asyncHandler(async(req, res) => {
-    const { _id } = req.body.user;
-    console.log("🚀 ~ file: cart.controller.js:46 ~ getUserCart ~ _id", _id)
+    const id = req.params.id;
+    if (!ObjectId.isValid(id)) throw new Error("Invalid id!");
     try {
-        const cart = await Cart.findOne({ orderBy: _id }).populate(
+        const cart = await Cart.findOne({ orderBy: id }).populate(
             "products.product"
         );
         console.log("🚀 ~ file: cart.controller.js:52 ~ getUserCart ~ cart", cart)
@@ -55,7 +58,8 @@ const getUserCart = asyncHandler(async(req, res) => {
 });
 
 const emptyCart = asyncHandler(async(req, res) => {
-    const { _id } = req.body.user;
+    const id = req.params.id;
+    if (!ObjectId.isValid(id)) throw new Error("Invalid id!");
     try {
         const user = await User.findOne({ _id });
         const cart = await Cart.findOneAndRemove({ orderBy: user._id });
